backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
redis
python-jose[cryptography]
openai
requests
python-dotenv
PyJWT
passlib[bcrypt]
websockets
google-genai
